move location prefix feature from api header to module local location
