+++ src/limestone/datastore_snapshot.cpp
@@ -19,6 +19,7 @@
 
 #include <glog/logging.h>
 #include <limestone/logging.h>
+#include "logging_helper.h"
 
 #include <limestone/api/datastore.h>
 #include "log_entry.h"
