+++ src/limestone/snapshot.cpp
@@ -17,6 +17,7 @@
 
 #include <glog/logging.h>
 #include <limestone/logging.h>
+#include "logging_helper.h"
 
 #include <boost/filesystem.hpp>
